# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
...
        "api/v1/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get current user's personal informaiton",
                "operationId": "getUser",
                "responses": {
                    "200": {
                        "description": "User personal info",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserDetailResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ControllerException"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/user": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Change user's personal information",
                "operationId": "editUser",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Name of the user",
                                        "type": "string"
                                    },
                                    "about": {
                                        "description": "Description of the user",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User's personal info",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserDetailResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ControllerException"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "ControllerException": {
                "title": "ControllerException",
                "description": "ControllerException model",
                "properties": {
                    "message": {
                        "title": "message",
                        "description": "Response message",
                        "type": "string"
                    },
                    "status": {
                        "title": "status",
                        "description": "Response status",
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ControllerException"
                }
            },
...
            "UserDetailResource": {
                "title": "UserDetailResource",
                "description": "UserDetailResource model",
                "properties": {
                    "data": {
                        "title": "data",
                        "description": "Response data",
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "string"
                                },
                                "name": {
                                    "type": "string"
                                },
                                "email": {
                                    "type": "string"
                                },
                                "about": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "message": {
                        "title": "message",
                        "description": "Response message",
                        "type": "string"
                    },
                    "status": {
                        "title": "status",
                        "description": "Response status",
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "UserDetailResource"
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://app.test/oauth/authorize",
                        "tokenUrl": "http://app.test/oauth/token",
                        "refreshUrl": "http://app.test/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}